WIP: fuse SubscriptionService and MemberService
